requirements.txt
mkdocs-macros-plugin==1.3.7
mkdocs-include-dir-to-nav==1.2.0
mkdocs-literate-nav==0.6.1
mkdocs-site-urls==0.2.0
mike==2.1.3
mkdocs-material ==9.6.4
mkdocs-jupyter==0.25.1
mkdocstrings==0.28.1
mkdocstrings-python==1.16.1
pymdown-extensions==10.14.3
pandas==2.2.3
Jinja2==3.1.5
mkdocs-material[extras]
copasi_basico==0.78
pydantic==2.9.2
pylint==3.3.1
pytest==8.3.3
pytest-asyncio==0.25.2
hydra-core==1.3.2
coverage==7.3.2
